Changed XML request from synchronous to async
You are currently opening your XML request as a synchronous request which is actually deprecated in chrome due to the fact it is too taxing on user PCs (having the request be synchronous means that the whole program stops until the request has resolved).  

As it stands chrome is currently treating your request as an asynchronous request anyway due to synchronous requests being deprecated and so all this change will do is clear your console of an annoying message.